adminUserController moved to AdminUserResource & small changes in adminUser view are made
