+++ AnimalWebApp/src/main/webapp/index.html
@@ -97,7 +97,7 @@
 				<ul class="dropdown-menu">
 					<li><a href="#ua/user/home/animals" ng-show="userRole == 'moderator'">{{ 'NAV_ANIMAL_LIST' | translate }}</a></li>
 					<li role="presentation" class="divider" ng-show="userRole == 'moderator'"></li>
-                    <li><a href="#ua/users/admin/users" ng-show="userRole == 'moderator'">{{ 'NAV_USER_LIST' | translate }}</a></li>
+                    <li><a href="#ua/user/admin/users" ng-show="userRole == 'moderator'">{{ 'NAV_USER_LIST' | translate }}</a></li>
                     <!--li role="presentation" class="divider" ng-show="userRole == 'лікар'"></li-->
                     <li><a href="#ua/user/doctor/animals" ng-show="userRole == 'doctor'">{{ 'NAV_ANIMAL_LIST_FOR_DOCTOR' | translate }}</a></li>
 				</ul>
